feat: import si-type information from A2
