+++ src/apps/Altinn.Register/test/Altinn.Register.IntegrationTests/PartyImport/A2PartyImportConsumerTests.cs
@@ -192,7 +192,7 @@ public class A2PartyImportConsumerTests
             {
                 UserId = userId,
                 UserUUID = partyUuid,
-                UserType = 2,
+                UserType = 1,
                 UserName = (string?)null,
                 ExternalIdentity = (string?)null,
                 IsReserved = false,
